Display each peep's timestamp alongside its text
The chitter-feed view has been modified to interpret the date_and_time_
created value associated with each peep, and display this value in a
more user-friendly format. If the peep was posted today, only the time
is shown, otherwise the day and month are also shown (a future feature
to consider adding is displaying the year selectively, depending on
whether the peep was posted this year/within the last 12 months).

For ease of reference in the code, and to improve separation of the
controller from the database for stubbing and adaptability,
ChitterHandler#retrieve_peeps now returns a hash of data for each peep.
Tests and views were updated accordingly. The query submitted to the
database now specifies that the records should be ordered by timestamp
in case this ceases to be the default behaviour in future.

Upper character limit on password length removed in docs/db-model as it
was needlessly restrictive.