Enable providing a username when posting a peep
An additional 'username' field has been added to both databases. This
was previously called 'user' in docs/db-model, but 'user' appears to be
a SQL reserved term and the field could not be added with that name.

As stated in the README, username currently has no function other than
allowing the peeper to personalise their message with a (non-unique)
tag of their choosing.