Implement creating a user account
Both databases now have a 'users' table to store account details.

Chitter will record the user as being signed in once the account is
created, but there is currently no other way to sign in/out.

Rubocop has raised 3 concerns over code quality which will be addressed
as a priority.