Rewrite test for peep metadata
Test was potentially giving false positives because the logged-in
username is displayed on the feed page. Two possible solutions:

1. Log the user out in the feature test before navigating to the feed
   page, and continue to search the whole `page` for the required content.
2. Make the feature test more specific - search for the required content
   within the peep by selecting it, by doing something like
   `expect(page.find('.peep')).to have_content`

Option 1 has the disadvantage that if at some point the text that you're
checking for is introduced anywhere else on the page then the test may
produce false positives. You can get around this by using something very
obscure for the name and username, like a string of random characters.

The date and time is more tricky though - it's conceivable that the
current date and time might be displayed elsewhere on the page at some
point, in which case expecting to find the date and time on the page may
give false positives.

Option 2 has the disadvantage that it's more tightly coupled to the
view layer - if at some point the `.peep` class is removed from the
peeps then the test will stop working.

So maybe it's best to use Option 2 for the date, and Option 1 for the
username and name. That way we minimise the tests which are coupled to
the view layer, and we use the more tightly coupled test only to test
the content string that we can't set to something we know will never be
on the page, i.e. the date.