Wrap peep database results in Peep class instances
This change required significant alterations to various classes.
The Peep.all method now returns an array of Peep objects rather
than just strings. This allows us to encapsulate all information
relating to a peep in a single object (so far only the text and
date it was created).

A create class method was also added to the Peep class. This
encapsulates the call to the database that adds the row to the
peeps table, and returns an instance of peep. This method is now
used instead of direct calls to the database in the various tests
that require peeps to be set up before they run.