Add peep author details to peeps list
Implementing this highlighted that the link between the peeps and
makers database tables was not working as expected.

An error was found in the all method of the Peep class that meant
the returned array of peeps did not have their Maker variable
correctly set. Once this was corrected, it was relatively
easy to add the details to the displayed list.