USER STORY 2
As a Maker
So that I can let people know what I am doing
I want to post a message (peep) to chitter

Created an option for the user to create a new peep that gets added to
the database. However, still need to refactor to the model and take SQL
queries away from the controller.