USER STORY:
As a Maker
So that I can better appreciate the context of a peep
I want to see the time at which it was made

We managed to join the timestamp to each peep(content) by iterating
within the model the different hash keys from within our table.

We initialized, calling instance variables equal to those hash keys from
within the class.

We then used those variables within the erb files to return the  '
'peep.content' and 'peep.time_stamp'.