Went back to a previous commit message,
Having created am extracted database_setup that merged into main, the
tests were passing, but the rackup local server was printing to the
screen the hash key in the format:

```
{:content => "This is a peep"}

Suspected that it is the creation of the extracted data_base_setup that
is causing this return as all other files (model, controller, view) are
exactly the same.

ONLY DIFFERENCE was 'database_connection_setup.rb'  AND
'database_connection.rb'

All RSpec passing.
Going to continue on this branch.